refactor(discount): move discount validation to a single method
